+++ src/main/java/bdv/img/hdf5/HDF5AccessHack.java
@@ -63,9 +63,9 @@ class HDF5AccessHack implements IHDF5Access
 {
 	private final IHDF5Reader hdf5Reader;
 
-	private final int fileId;
+	private final long fileId;
 
-	private final int numericConversionXferPropertyListID;
+	private final long numericConversionXferPropertyListID;
 
 	private final long[] reorderedDimensions = new long[ 3 ];
 
@@ -139,7 +139,7 @@ class HDF5AccessHack implements IHDF5Access
 		final Class< ? > k2 = Class.forName( "ch.systemsx.cisd.hdf5.HDF5BaseReader" );
 		final Field f2 = k2.getDeclaredField( "fileId" );
 		f2.setAccessible( true );
-		fileId = ( ( Integer ) f2.get( baseReader ) ).intValue();
+		fileId = ( ( Long ) f2.get( baseReader ) ).longValue();
 
 		final Field f3 = k2.getDeclaredField( "h5" );
 		f3.setAccessible( true );
@@ -148,7 +148,7 @@ class HDF5AccessHack implements IHDF5Access
 		final Class< ? > k4 = Class.forName( "ch.systemsx.cisd.hdf5.HDF5" );
 		final Field f4 = k4.getDeclaredField( "numericConversionXferPropertyListID" );
 		f4.setAccessible( true );
-		numericConversionXferPropertyListID = ( ( Integer ) f4.get( h5 ) ).intValue();
+		numericConversionXferPropertyListID = ( ( Long ) f4.get( h5 ) ).longValue();
 
 		openDataSetCache = new OpenDataSetCache();
 	}
